kafka-pub-sub/python-publisher/requirements.txt
essential_generators
kafka-python